@@ -1,18 +1,24 @@
 {
   "name": "@inotom/expand-controller",
-  "version": "1.0.0",
+  "version": "2.0.0",
   "description": "Toggle expand state script.",
-  "main": "dist/index.js",
-  "cdn": "dist/expand-controller.min.js",
-  "config": {
-    "dist_dir": "dist",
-    "src_dir": "src"
-  },
+  "main": "dist/index.cjs.js",
+  "module": "dist/index.esm.js",
+  "unpkg": "dist/expand-controller.js",
+  "files": [
+    "dist"
+  ],
+  "types": "dist/index.d.ts",
   "scripts": {
-    "start": "NODE_ENV=development npx rollup -c -w",
-    "build": "NODE_ENV=production npx rollup -c",
-    "lint": "eslint --ext .js $npm_package_config_src_dir",
-    "prettier": "prettier --write $npm_package_config_src_dir/**/*.js",
+    "start": "NODE_ENV=development rollup -c -w",
+    "prebuild": "rimraf dist/*.js dist/*.map dist/*.d.ts && npm run clean && npm run types",
+    "build": "NODE_ENV=production rollup -c",
+    "clean": "rimraf '**/.DS_Store'",
+    "prepublishOnly": "npm run build",
+    "release": "npm publish",
+    "types": "tsc --emitDeclarationOnly",
+    "lint": "eslint --ext .ts src",
+    "prettier": "prettier --write src/**/*.ts",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
@@ -26,17 +32,24 @@
     "url": "https://github.com/inotom/expand-controller/issues"
   },
   "homepage": "https://github.com/inotom/expand-controller#readme",
+  "publishConfig": {
+    "access": "public"
+  },
   "devDependencies": {
-    "@babel/core": "7.7.7",
-    "@babel/plugin-proposal-optional-chaining": "7.7.5",
-    "@babel/preset-env": "7.7.7",
-    "babel-eslint": "10.0.3",
-    "core-js": "3.6.1",
-    "eslint": "6.8.0",
-    "prettier": "1.19.1",
-    "rollup": "1.27.14",
-    "rollup-plugin-babel": "4.3.3",
-    "rollup-plugin-license": "0.13.0",
-    "rollup-plugin-terser": "5.1.3"
+    "@rollup/plugin-buble": "0.21.3",
+    "@rollup/plugin-commonjs": "14.0.0",
+    "@rollup/plugin-node-resolve": "8.4.0",
+    "@types/node": "14.0.26",
+    "@typescript-eslint/eslint-plugin": "3.7.0",
+    "@typescript-eslint/parser": "3.7.0",
+    "eslint": "7.5.0",
+    "lodash.camelcase": "4.3.0",
+    "lodash.upperfirst": "4.3.1",
+    "prettier": "2.0.5",
+    "rimraf": "3.0.2",
+    "rollup": "2.23.0",
+    "rollup-plugin-terser": "6.1.0",
+    "rollup-plugin-typescript2": "0.27.1",
+    "typescript": "3.9.7"
   }
 }
