@@ -1,43 +1,69 @@
 /** @prettier */
 
-import babel from 'rollup-plugin-babel';
+import resolve from '@rollup/plugin-node-resolve';
+import typescript from 'rollup-plugin-typescript2';
+import commonjs from '@rollup/plugin-commonjs';
+import buble from '@rollup/plugin-buble';
 import { terser } from 'rollup-plugin-terser';
-import license from 'rollup-plugin-license';
 
-const pkg = require('./package.json');
+import camelCase from 'lodash.camelcase';
+import upperFirst from 'lodash.upperfirst';
 
-const plugins = [
-  babel({
-    exclude: 'node_modules/**',
-    runtimeHelpers: true,
-    babelrc: true,
-  }),
-  terser({
-    ecma: 5,
-  }),
-  license({
-    banner: `${pkg.name} v${pkg.version} ${pkg.author} | ${pkg.license}`,
-  }),
+import pkg from './package.json';
+
+// Remove npm package scope.
+const unscopedName = pkg.name.replace(/^\@.*\//, '');
+const moduleName = upperFirst(camelCase(unscopedName));
+
+// ライブラリに埋め込むcopyright
+const banner = `/*! ${unscopedName}.js v${pkg.version} ${pkg.author} | ${pkg.license} License */`;
+
+export default [
+  // Options for browser.
+  {
+    input: 'src/index.ts',
+    output: [
+      {
+        name: moduleName,
+        file: pkg.unpkg,
+        format: 'iife',
+        sourcemap: 'inline',
+        banner,
+      },
+      // minify
+      {
+        name: moduleName,
+        file: pkg.unpkg.replace('.js', '.min.js'),
+        format: 'iife',
+        banner,
+        plugins: [terser()],
+      },
+    ],
+    // Exclude development modules.
+    external: [...Object.keys(pkg.devDependencies || {})],
+    plugins: [resolve(), typescript(), commonjs({ extensions: ['.ts', '.js'] }), buble()],
+  },
+  // Options for module.
+  {
+    input: 'src/index.ts',
+    output: [
+      // CommonJS
+      {
+        file: pkg.main,
+        format: 'cjs',
+        sourcemap: true,
+        banner,
+      },
+      // ES Modules
+      {
+        file: pkg.module,
+        format: 'es',
+        sourcemap: true,
+        banner,
+      },
+    ],
+    // Exclude other modules.
+    external: [...Object.keys(pkg.dependencies || {}), ...Object.keys(pkg.devDependencies || {})],
+    plugins: [resolve(), typescript(), commonjs({ extensions: ['.ts', '.js'] })],
+  },
 ];
-
-export default {
-  input: './src/index.js',
-  output: [
-    {
-      file: './dist/index.js',
-      format: 'cjs',
-    },
-    {
-      file: './dist/expand-controller.min.js',
-      format: 'iife',
-      name: 'ExpandController',
-    },
-  ],
-  external: [
-    'core-js/modules/es.array.slice',
-    'core-js/modules/es.regexp.exec',
-    'core-js/modules/es.string.split',
-    'core-js/modules/web.dom-collections.for-each',
-  ],
-  plugins,
-};
